Cleanup filter building API
