+++ packages/components/src/test/instance-filter-builder/IntanceFilterBuilder.test.tsx
@@ -18,25 +18,15 @@ import { EmptyLocalization } from "@itwin/core-common";
 import { FormattingUnitSystemChangedArgs, IModelApp, IModelConnection } from "@itwin/core-frontend";
 import { FormatterSpec, ParserSpec } from "@itwin/core-quantity";
 import { SchemaContext } from "@itwin/ecschema-metadata";
-import { ClassInfo, Descriptor, KoqPropertyValueFormatter, NavigationPropertyInfo } from "@itwin/presentation-common";
+import { ClassInfo, Descriptor, KoqPropertyValueFormatter } from "@itwin/presentation-common";
 import { Presentation } from "@itwin/presentation-frontend";
 import { fireEvent, render, waitFor } from "@testing-library/react";
 import { renderHook } from "@testing-library/react-hooks";
 import { SchemaMetadataContextProvider } from "../../presentation-components/common/SchemaMetadataContext";
 import { ECClassInfo, getIModelMetadataProvider } from "../../presentation-components/instance-filter-builder/ECMetadataProvider";
-import {
-  InstanceFilterBuilder,
-  useFilterBuilderNavigationPropertyEditorContext,
-  usePresentationInstanceFilteringProps,
-} from "../../presentation-components/instance-filter-builder/InstanceFilterBuilder";
-import { INSTANCE_FILTER_FIELD_SEPARATOR } from "../../presentation-components/instance-filter-builder/Utils";
+import { InstanceFilterBuilder, usePresentationInstanceFilteringProps } from "../../presentation-components/instance-filter-builder/InstanceFilterBuilder";
 import { createTestECClassInfo, stubRaf } from "../_helpers/Common";
-import {
-  createTestCategoryDescription,
-  createTestContentDescriptor,
-  createTestPropertiesContentField,
-  createTestSimpleContentField,
-} from "../_helpers/Content";
+import { createTestCategoryDescription, createTestContentDescriptor, createTestPropertiesContentField } from "../_helpers/Content";
 
 describe("InstanceFilterBuilder", () => {
   stubRaf();
@@ -451,71 +441,5 @@ describe("usePresentationInstanceFilteringProps", () => {
       result.current.onRulePropertySelected({ name: "invalidProp", displayLabel: "InvalidProp", typename: "string" });
       expect(result.current.selectedClasses).to.be.empty;
     });
-  });
-});
-
-describe("useFilterBuilderNavigationPropertyEditorContext", () => {
-  interface Props {
-    imodel: IModelConnection;
-    descriptor: Descriptor;
-  }
-  const testImodel = {} as IModelConnection;
-
-  it("returns navigation property info", async () => {
-    const navigationPropertyInfo: NavigationPropertyInfo = {
-      classInfo: { id: "2", label: "Prop Class", name: "TestSchema:PropClass" },
-      targetClassInfo: { id: "3", label: "Target Class", name: "TestSchema:TargetClass" },
-      isForwardRelationship: true,
-      isTargetPolymorphic: true,
-    };
-    const fieldName = "field_name";
-    const testDescriptor = createTestContentDescriptor({
-      fields: [
-        createTestPropertiesContentField({
-          name: fieldName,
-          properties: [
-            {
-              property: {
-                classInfo: { id: "1", label: "Field Class", name: "TestSchema:FieldClass" },
-                name: "nav_prop",
-                type: "navigation",
-                navigationPropertyInfo,
-              },
-            },
-          ],
-        }),
-      ],
-    });
-    const propertyDescription: PropertyDescription = {
-      displayLabel: "TestProp",
-      name: `test_category${INSTANCE_FILTER_FIELD_SEPARATOR}${fieldName}`,
-      typename: "navigation",
-    };
-
-    const { result } = renderHook(({ imodel, descriptor }: Props) => useFilterBuilderNavigationPropertyEditorContext(imodel, descriptor), {
-      initialProps: { imodel: testImodel, descriptor: testDescriptor },
-    });
-
-    const info = await result.current.getNavigationPropertyInfo(propertyDescription);
-    expect(info).to.be.deep.eq(navigationPropertyInfo);
-  });
-
-  it("returns `undefined` for non properties field", async () => {
-    const fieldName = "field_name";
-    const testDescriptor = createTestContentDescriptor({
-      fields: [createTestSimpleContentField({ name: fieldName })],
-    });
-    const propertyDescription: PropertyDescription = {
-      displayLabel: "TestProp",
-      name: `test_category${INSTANCE_FILTER_FIELD_SEPARATOR}${fieldName}`,
-      typename: "navigation",
-    };
-
-    const { result } = renderHook(({ imodel, descriptor }: Props) => useFilterBuilderNavigationPropertyEditorContext(imodel, descriptor), {
-      initialProps: { imodel: testImodel, descriptor: testDescriptor },
-    });
-
-    const info = await result.current.getNavigationPropertyInfo(propertyDescription);
-    expect(info).to.be.undefined;
   });
 });
