+++ packages/components/src/test/instance-filter-builder/PresentationInstanceFilterDialog.test.tsx
@@ -15,9 +15,9 @@ import { Descriptor } from "@itwin/presentation-common";
 import { Presentation } from "@itwin/presentation-frontend";
 import { waitFor } from "@testing-library/react";
 import { ECClassInfo, getIModelMetadataProvider } from "../../presentation-components/instance-filter-builder/ECMetadataProvider";
+import * as presentationFilterBuilder from "../../presentation-components/instance-filter-builder/PresentationFilterBuilder";
 import { PresentationInstanceFilterDialog } from "../../presentation-components/instance-filter-builder/PresentationInstanceFilterDialog";
 import { PresentationInstanceFilterInfo } from "../../presentation-components/instance-filter-builder/Types";
-import * as instanceFilterBuilderUtils from "../../presentation-components/instance-filter-builder/Utils";
 import { createTestECClassInfo, render, stubDOMMatrix, stubRaf } from "../_helpers/Common";
 import { createTestCategoryDescription, createTestContentDescriptor, createTestPropertiesContentField } from "../_helpers/Content";
 
@@ -135,7 +135,7 @@ describe("PresentationInstanceFilterDialog", () => {
   });
 
   it("does not invoke `onApply` when filter is missing presentation metadata", async () => {
-    sinon.stub(instanceFilterBuilderUtils, "createPresentationInstanceFilter").returns(undefined);
+    sinon.stub(presentationFilterBuilder, "createPresentationInstanceFilter").returns(undefined);
     const spy = sinon.spy();
     const { container, getByText, user } = render(
       <PresentationInstanceFilterDialog imodel={imodelMock.object} descriptor={descriptor} onClose={() => {}} onApply={spy} isOpen={true} />,
