+++ packages/components/src/test/instance-filter-builder/Utils.test.ts
@@ -6,24 +6,18 @@
 import { expect } from "chai";
 import sinon from "sinon";
 import { PropertyDescription, PropertyValueFormat, StandardTypeNames } from "@itwin/appui-abstract";
-import {
-  PropertyFilter,
-  PropertyFilterRule,
-  PropertyFilterRuleGroup,
-  PropertyFilterRuleGroupOperator,
-  PropertyFilterRuleOperator,
-} from "@itwin/components-react";
+import { PropertyFilterRuleOperator } from "@itwin/components-react";
 import { EmptyLocalization } from "@itwin/core-common";
-import { Field } from "@itwin/presentation-common";
+import { IModelConnection } from "@itwin/core-frontend";
+import { Descriptor, NavigationPropertyInfo } from "@itwin/presentation-common";
 import { Presentation } from "@itwin/presentation-frontend";
-import { PresentationInstanceFilter } from "../../presentation-components/instance-filter-builder/Types";
+import { renderHook } from "@testing-library/react-hooks";
 import {
-  convertPresentationFilterToPropertyFilter,
   createInstanceFilterPropertyInfos,
-  createPresentationInstanceFilter,
   DEFAULT_ROOT_CATEGORY_NAME,
   filterRuleValidator,
   INSTANCE_FILTER_FIELD_SEPARATOR,
+  useFilterBuilderNavigationPropertyEditorContext,
 } from "../../presentation-components/instance-filter-builder/Utils";
 import { createTestECClassInfo } from "../_helpers/Common";
 import {
@@ -31,11 +25,8 @@ import {
   createTestContentDescriptor,
   createTestNestedContentField,
   createTestPropertiesContentField,
+  createTestSimpleContentField,
 } from "../_helpers/Content";
-
-function getPropertyDescriptionName(field: Field) {
-  return `root${INSTANCE_FILTER_FIELD_SEPARATOR}${field.name}`;
-}
 
 describe("createInstanceFilterPropertyInfos", () => {
   it("creates property infos when fields are in root category", () => {
@@ -143,245 +134,6 @@ describe("createInstanceFilterPropertyInfos", () => {
   });
 });
 
-describe("createPresentationInstanceFilter", () => {
-  const category = createTestCategoryDescription({ name: "root", label: "Root" });
-  const propertyField1 = createTestPropertiesContentField({
-    properties: [{ property: { classInfo: createTestECClassInfo(), name: "prop1", type: "string" } }],
-    category,
-    name: "propField1",
-  });
-  const propertyField2 = createTestPropertiesContentField({
-    properties: [{ property: { classInfo: createTestECClassInfo(), name: "prop1", type: "string" } }],
-    category,
-    name: "propField1",
-  });
-  const descriptor = createTestContentDescriptor({
-    categories: [category],
-    fields: [propertyField1, propertyField2],
-  });
-
-  it("finds properties fields for property description", () => {
-    const filter: PropertyFilterRuleGroup = {
-      operator: PropertyFilterRuleGroupOperator.And,
-      rules: [
-        {
-          property: { name: getPropertyDescriptionName(propertyField1), displayLabel: "Prop1", typename: "string" },
-          operator: PropertyFilterRuleOperator.IsNull,
-        },
-        {
-          property: { name: getPropertyDescriptionName(propertyField2), displayLabel: "Prop2", typename: "string" },
-          operator: PropertyFilterRuleOperator.IsNull,
-        },
-      ],
-    };
-    expect(createPresentationInstanceFilter(descriptor, filter)).to.matchSnapshot();
-  });
-
-  it("returns filter condition when group has only one rule", () => {
-    const filter: PropertyFilterRuleGroup = {
-      operator: PropertyFilterRuleGroupOperator.And,
-      rules: [
-        {
-          property: { name: getPropertyDescriptionName(propertyField1), displayLabel: "Prop1", typename: "string" },
-          operator: PropertyFilterRuleOperator.IsNull,
-        },
-      ],
-    };
-    expect(createPresentationInstanceFilter(descriptor, filter)).to.containSubset({
-      operator: PropertyFilterRuleOperator.IsNull,
-      field: propertyField1,
-    });
-  });
-
-  it("returns undefined if filter group is empty", () => {
-    expect(createPresentationInstanceFilter(descriptor, { operator: PropertyFilterRuleGroupOperator.And, rules: [] })).to.be.undefined;
-  });
-
-  it("returns undefined when rule properties field cannot be found", () => {
-    const property: PropertyDescription = { name: `${INSTANCE_FILTER_FIELD_SEPARATOR}invalidFieldName`, displayLabel: "Prop", typename: "string" };
-    expect(createPresentationInstanceFilter(descriptor, { property, operator: PropertyFilterRuleOperator.IsNull })).to.be.undefined;
-  });
-
-  it("returns undefined when group has rule with invalid property field", () => {
-    const filter: PropertyFilterRuleGroup = {
-      operator: PropertyFilterRuleGroupOperator.And,
-      rules: [
-        {
-          property: { name: getPropertyDescriptionName(propertyField1), displayLabel: "Prop1", typename: "string" },
-          operator: PropertyFilterRuleOperator.IsNull,
-        },
-        {
-          property: { name: `${INSTANCE_FILTER_FIELD_SEPARATOR}invalidFieldName`, displayLabel: "Prop2", typename: "string" },
-          operator: PropertyFilterRuleOperator.IsNull,
-        },
-      ],
-    };
-    expect(createPresentationInstanceFilter(descriptor, filter)).to.be.undefined;
-  });
-
-  it("returns undefined when rule has non primitive value", () => {
-    const filter: PropertyFilterRule = {
-      property: { name: getPropertyDescriptionName(propertyField1), displayLabel: "Prop1", typename: "string" },
-      operator: PropertyFilterRuleOperator.IsEqual,
-      value: { valueFormat: PropertyValueFormat.Array, items: [], itemsTypeName: "number" },
-    };
-    expect(createPresentationInstanceFilter(descriptor, filter)).to.be.undefined;
-  });
-});
-
-describe("convertPresentationInstanceFilterToInstanceFilter", () => {
-  const category = createTestCategoryDescription({ name: "root", label: "Root" });
-  const propertyField1 = createTestPropertiesContentField({
-    properties: [{ property: { classInfo: createTestECClassInfo(), name: "prop1", type: "string" } }],
-    category,
-    name: "propField1",
-    label: "Prop1",
-  });
-  const propertyField2 = createTestPropertiesContentField({
-    properties: [{ property: { classInfo: createTestECClassInfo(), name: "prop2", type: "string" } }],
-    category,
-    name: "propField2",
-    label: "Prop2",
-  });
-  const propertyField3 = createTestPropertiesContentField({
-    properties: [{ property: { classInfo: createTestECClassInfo(), name: "prop3", type: "string" } }],
-    category,
-    name: "propField3",
-    label: "Prop3",
-  });
-  const nestedField = createTestNestedContentField({
-    nestedFields: [propertyField3],
-    category,
-    name: "nestedField",
-    label: "NestedProp",
-  });
-  const nestedField2 = createTestNestedContentField({
-    nestedFields: [nestedField],
-    category,
-    name: "nestedField2",
-    label: "NestedProp2",
-  });
-  propertyField3.rebuildParentship(nestedField);
-  const descriptor = createTestContentDescriptor({
-    categories: [category],
-    fields: [propertyField1, propertyField2, nestedField2],
-  });
-
-  it("property filter converts to presentation filter and vise versa correctly", () => {
-    const filter: PropertyFilter = {
-      operator: PropertyFilterRuleGroupOperator.And,
-      rules: [
-        {
-          property: { name: getPropertyDescriptionName(propertyField1), displayLabel: "Prop1", typename: "string" },
-          operator: PropertyFilterRuleOperator.IsNull,
-          value: undefined,
-        },
-        {
-          property: { name: getPropertyDescriptionName(propertyField2), displayLabel: "Prop2", typename: "string" },
-          operator: PropertyFilterRuleOperator.IsNull,
-          value: undefined,
-        },
-      ],
-    };
-
-    const presentationFilter = createPresentationInstanceFilter(descriptor, filter);
-    expect(presentationFilter).to.not.be.undefined;
-    const result = presentationFilter ? convertPresentationFilterToPropertyFilter(descriptor, presentationFilter) : undefined;
-    expect(result).to.be.deep.eq(filter);
-  });
-
-  it("converts presentation filter with nested conditions to property filter", () => {
-    const presentationFilter: PresentationInstanceFilter = {
-      operator: PropertyFilterRuleGroupOperator.And,
-      conditions: [
-        {
-          operator: PropertyFilterRuleGroupOperator.And,
-          conditions: [
-            {
-              field: propertyField1,
-              operator: PropertyFilterRuleOperator.IsNull,
-              value: undefined,
-            },
-          ],
-        },
-      ],
-    };
-
-    const propertyFilter: PropertyFilter = {
-      operator: PropertyFilterRuleGroupOperator.And,
-      rules: [
-        {
-          operator: PropertyFilterRuleGroupOperator.And,
-          rules: [
-            {
-              property: { name: getPropertyDescriptionName(propertyField1), displayLabel: "Prop1", typename: "string" },
-              operator: PropertyFilterRuleOperator.IsNull,
-              value: undefined,
-            },
-          ],
-        },
-      ],
-    };
-
-    const result = convertPresentationFilterToPropertyFilter(descriptor, presentationFilter);
-    expect(result).to.be.deep.eq(propertyFilter);
-  });
-
-  it("converts presentation filter with nested fields to property filter", () => {
-    const presentationFilter: PresentationInstanceFilter = {
-      operator: PropertyFilterRuleGroupOperator.And,
-      conditions: [
-        {
-          field: propertyField3,
-          operator: PropertyFilterRuleOperator.IsNull,
-          value: undefined,
-        },
-      ],
-    };
-
-    const propertyFilter: PropertyFilter = {
-      operator: PropertyFilterRuleGroupOperator.And,
-      rules: [
-        {
-          property: {
-            name: `${getPropertyDescriptionName(nestedField2)}$${nestedField.name}$${propertyField3.name}`,
-            displayLabel: "Prop3",
-            typename: "string",
-          },
-          operator: PropertyFilterRuleOperator.IsNull,
-          value: undefined,
-        },
-      ],
-    };
-
-    const result = convertPresentationFilterToPropertyFilter(descriptor, presentationFilter);
-    expect(result).to.be.deep.eq(propertyFilter);
-  });
-
-  it("returns undefined if property used in filter is not found in descriptor", () => {
-    const propertyField = createTestPropertiesContentField({
-      properties: [{ property: { classInfo: createTestECClassInfo(), name: "prop", type: "string" } }],
-      category,
-      name: "propField",
-      label: "Prop",
-    });
-
-    const presentationFilter: PresentationInstanceFilter = {
-      operator: PropertyFilterRuleGroupOperator.And,
-      conditions: [
-        {
-          field: propertyField,
-          operator: PropertyFilterRuleOperator.IsNull,
-          value: undefined,
-        },
-      ],
-    };
-
-    const result = convertPresentationFilterToPropertyFilter(descriptor, presentationFilter);
-    expect(result).to.be.undefined;
-  });
-});
-
 describe("filterRuleValidator", () => {
   const numericProperty: PropertyDescription = {
     displayLabel: "Numeric Prop",
@@ -465,5 +217,71 @@ describe("filterRuleValidator", () => {
         },
       }),
     ).to.be.undefined;
+  });
+});
+
+describe("useFilterBuilderNavigationPropertyEditorContext", () => {
+  interface Props {
+    imodel: IModelConnection;
+    descriptor: Descriptor;
+  }
+  const testImodel = {} as IModelConnection;
+
+  it("returns navigation property info", async () => {
+    const navigationPropertyInfo: NavigationPropertyInfo = {
+      classInfo: { id: "2", label: "Prop Class", name: "TestSchema:PropClass" },
+      targetClassInfo: { id: "3", label: "Target Class", name: "TestSchema:TargetClass" },
+      isForwardRelationship: true,
+      isTargetPolymorphic: true,
+    };
+    const fieldName = "field_name";
+    const testDescriptor = createTestContentDescriptor({
+      fields: [
+        createTestPropertiesContentField({
+          name: fieldName,
+          properties: [
+            {
+              property: {
+                classInfo: { id: "1", label: "Field Class", name: "TestSchema:FieldClass" },
+                name: "nav_prop",
+                type: "navigation",
+                navigationPropertyInfo,
+              },
+            },
+          ],
+        }),
+      ],
+    });
+    const propertyDescription: PropertyDescription = {
+      displayLabel: "TestProp",
+      name: `test_category${INSTANCE_FILTER_FIELD_SEPARATOR}${fieldName}`,
+      typename: "navigation",
+    };
+
+    const { result } = renderHook(({ imodel, descriptor }: Props) => useFilterBuilderNavigationPropertyEditorContext(imodel, descriptor), {
+      initialProps: { imodel: testImodel, descriptor: testDescriptor },
+    });
+
+    const info = await result.current.getNavigationPropertyInfo(propertyDescription);
+    expect(info).to.be.deep.eq(navigationPropertyInfo);
+  });
+
+  it("returns `undefined` for non properties field", async () => {
+    const fieldName = "field_name";
+    const testDescriptor = createTestContentDescriptor({
+      fields: [createTestSimpleContentField({ name: fieldName })],
+    });
+    const propertyDescription: PropertyDescription = {
+      displayLabel: "TestProp",
+      name: `test_category${INSTANCE_FILTER_FIELD_SEPARATOR}${fieldName}`,
+      typename: "navigation",
+    };
+
+    const { result } = renderHook(({ imodel, descriptor }: Props) => useFilterBuilderNavigationPropertyEditorContext(imodel, descriptor), {
+      initialProps: { imodel: testImodel, descriptor: testDescriptor },
+    });
+
+    const info = await result.current.getNavigationPropertyInfo(propertyDescription);
+    expect(info).to.be.undefined;
   });
 });
